requirements.txt
groq
langgraph
gspread
google-auth
python-telegram-bot
python-dotenv
